Validando ação

+++ src/users/users.service.ts
@@ -3,6 +3,7 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDto } from './dto/create.user.dto';
 import { UpdateUserDto } from './dto/update.user.dto';
 import { HashingServiceProtocol } from 'src/auth/hash/hashing.service';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class UsersService {
@@ -43,7 +44,11 @@ export class UsersService {
     }
   }
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
+  async update(
+    id: number,
+    updateUserDto: UpdateUserDto,
+    tokenPayload: PayloadTokenDto,
+  ) {
     try {
       const user = await this.prisma.user.findFirst({
         where: { id },
@@ -51,6 +56,9 @@ export class UsersService {
 
       if (!user)
         throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
+
+      if (user.id !== tokenPayload.sub)
+        throw new HttpException('Acesso negado', HttpStatus.BAD_REQUEST);
 
       const dataUser: {
         name?: string;
@@ -84,7 +92,7 @@ export class UsersService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, tokenPayload: PayloadTokenDto) {
     try {
       const user = await this.prisma.user.findFirst({
         where: { id },
@@ -92,6 +100,9 @@ export class UsersService {
 
       if (!user)
         throw new HttpException('User not found', HttpStatus.BAD_REQUEST);
+
+      if (user.id !== tokenPayload.sub)
+        throw new HttpException('Acesso negado', HttpStatus.BAD_REQUEST);
 
       const userDelete = await this.prisma.user.delete({
         where: { id: user.id },
